* feat(settings): introduce minimum SNR threshold for LoRa peers and im…

* feat(mesh): improve LoRa timing, add flux guard, and enhance retry ha…

* fix(stability): overflow-safe timers, buffer bounds checks, duty cycl…

* Merge remote-tracking branch 'origin/dev-next' into webui-rework

+++ src/helperFunctions.h
@@ -20,7 +20,7 @@ struct FileWriteParams {
 
 void printHexArray(uint8_t* data, size_t length);
 void addJSONtoFile(char* buffer, size_t length, const char* file, const uint16_t lines);
-uint32_t getTOA(uint8_t payloadBytes);
+uint32_t getTOA(uint16_t payloadBytes);
 void sendMessage(const char* dst, const char* text, uint8_t messageType = Frame::MessageTypes::TEXT_MESSAGE); 
 void sendGroup(const char* dst, const char* text, uint8_t messageType = Frame::MessageTypes::TEXT_MESSAGE); 
 void safeUtf8Copy(char* dest, const uint8_t* src, size_t maxLength);
